+++ ChangeLog.md
@@ -24,6 +24,8 @@ See docs/process.md for more on how version tagging works.
   static string data in your program.   This has long been part of the native
   ELF linker and should not be observable in well-behaved programs.  This
   behavior can be disabled by passing `-Wl,-O0`.
+- Reinstated the warning on linker-only `-s` settings passed when not linking
+  (i.e.  when compiling with `-c`).
 
 2.0.20: 05/04/2021
 ------------------